move slashing to latest rpc patterns
