+++ api/client/client.go
@@ -9,7 +9,7 @@ import (
 	healthRpc "github.com/textileio/powergate/health/rpc"
 	askRpc "github.com/textileio/powergate/index/ask/rpc"
 	minerPb "github.com/textileio/powergate/index/miner/pb"
-	slashingPb "github.com/textileio/powergate/index/slashing/pb"
+	slashingRpc "github.com/textileio/powergate/index/slashing/rpc"
 	netRpc "github.com/textileio/powergate/net/rpc"
 	reputationRpc "github.com/textileio/powergate/reputation/rpc"
 	"github.com/textileio/powergate/util"
@@ -69,7 +69,7 @@ func NewClient(ma multiaddr.Multiaddr, opts ...grpc.DialOption) (*Client, error)
 	client := &Client{
 		Asks:       &Asks{client: askRpc.NewAPIClient(conn)},
 		Miners:     &Miners{client: minerPb.NewAPIClient(conn)},
-		Slashing:   &Slashing{client: slashingPb.NewAPIClient(conn)},
+		Slashing:   &Slashing{client: slashingRpc.NewAPIClient(conn)},
 		Deals:      &Deals{client: dealsRpc.NewAPIClient(conn)},
 		Wallet:     &Wallet{client: walletRpc.NewAPIClient(conn)},
 		Reputation: &Reputation{client: reputationRpc.NewAPIClient(conn)},
